Clean up srpv3 tests.
